Improve default subPatches
The default subPatches now only apply if the relevant plugin is inside
the users `plugin` list. This means that the user won't have to pay for
what they don't use.

Additionally, the functions to create custom patches are now available
in `patchUtils.nix`. This means that it'll be much easier for users to
add their own patches.