Cache string lookups and keep track of the string types
With this commit stirng lookups are now cached. This was needed to
remember the 'string type' of the string, whether it was single quote,
double quote or multiline. See the 2 added tests.

The easiest way in my mind to do this is to just cache the current
string start and end, and from the start also cache the type of string.
So that's what I did.

One downside of this approach is that for the cache to work, every
method now requires *Self instead of Self. This might not be needed
everywhere, but I'm a lazy motherfucker so I just did a search and
replace over the entire file. Might improve this in the future.