Stop copying over neovim-unwrapped
Code from
https://github.com/BirdeeHub/nixCats-nvim/commit/9f1bd3e4ce4b84576a7e3dce2b86d31f38e22648

The main idea is to not copy over the neovim-unwrapped files, as we
don't use most of them anyway. This allows us to properly alias to
`nvim` as the nvim binary is nowhere in nv.

We also have a slightly more minimal build, as nothing happens in the
buildPhase anymore